// quantity-limiter-backend/src/shared/api/types/wix-api/wix-api.enum.ts
export enum ThemeRole {
  Unpublished = 'unpublished',
  Main = 'main',
  Demo = 'demo',
  Development = 'development',
}

export enum PricingPlan {
  Free = 'basic',
  Standard = 'standard',
  Pro = 'pro',
  Plus = 'plus',
}

export enum PlanSubscription {
  Monthly = 'monthly',
  Yearly = 'yearly',
}